+ user-specific data. Broken query params for memes page

+++ app/api/categories/[id]/route.ts
@@ -1,16 +1,21 @@
 import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/prisma/client";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../../auth/authOptions";
 
 export async function DELETE(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session?.user?.email)
+    return NextResponse.json("Not authenticated", { status: 401 });
+
   const cat = await prisma.category.findUnique({
     where: {
       id: parseInt(params.id),
     },
   });
-
   if (!cat) {
     return NextResponse.json(
       { error: "Category does not exist." },
@@ -19,9 +24,18 @@ export async function DELETE(
   }
 
   try {
+    const user = await prisma.user.findUnique({
+      where: {
+        email: session!.user!.email!,
+      },
+    });
+    if (!user) {
+      return NextResponse.json("User not found", { status: 404 });
+    }
     const deleted = await prisma.category.delete({
       where: {
         id: cat.id,
+        authorId: user.id,
       },
     });
     return NextResponse.json(deleted);
@@ -37,15 +51,29 @@ export async function PUT(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
-  const { name, authorId } = await request.json();
+  const session = await getServerSession(authOptions);
+  if (!session?.user?.email)
+    return NextResponse.json("Not authenticated", { status: 401 });
 
-  if (!name || !authorId) {
-    return NextResponse.json("Error: Category data undefined.", { status: 400 });
+  const { name } = await request.json();
+
+  if (!name) {
+    return NextResponse.json("Error: Category data undefined.", {
+      status: 400,
+    });
   }
 
   try {
+    const user = await prisma.user.findUnique({
+      where: {
+        email: session!.user!.email!,
+      },
+    });
+    if (!user) {
+      return NextResponse.json("User not found", { status: 404 });
+    }
     const updatedCategory = await prisma.category.update({
-      where: { id: parseInt(params.id) },
+      where: { id: parseInt(params.id), authorId: user.id },
       data: {
         ...(name && { name }),
       },
@@ -64,12 +92,33 @@ export async function GET(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
-  const cat = await prisma.category.findUnique({
-    where: { id: parseInt(params.id) },
-  });
+  const session = await getServerSession(authOptions);
+  if (!session?.user?.email)
+    return NextResponse.json("Not authenticated", { status: 401 });
 
-  if (!cat) {
-    return NextResponse.json("Category not found.", { status: 404 });
+  try {
+    const user = await prisma.user.findUnique({
+      where: {
+        email: session!.user!.email!,
+      },
+    });
+    if (!user) {
+      return NextResponse.json("User not found", { status: 404 });
+    }
+    const cat = await prisma.category.findUnique({
+      where: {
+        id: parseInt(params.id),
+        authorId: user.id,
+      },
+    });
+
+    if (!cat) {
+      return NextResponse.json("Category not found.", { status: 404 });
+    }
+    return NextResponse.json(cat);
+  } catch {
+    return NextResponse.json("Could not get Category. Server error.", {
+      status: 500,
+    });
   }
-  return NextResponse.json(cat);
 }
